refactor(clean): estrarre read_sql_utils.py e read_excel.py
Separa gli helper SQL e l'adapter Excel da duckdb_read.py:

read_sql_utils.py:
- q_ident, sql_path, quote_list, csv_trim_projection

read_excel.py:
- _normalize_excel_sheet_name, _trim_excel_dataframe,
  _load_excel_frame, _execute_excel_read
- duckdb_read.pywrappa il risultato in ReadInfo per evitare
  circular import (read_excel non può importare da duckdb_read)

Nessun contract esterno cambiato; backward compat preservata.
25/25 test passati.
