+++ toolkit/clean/duckdb_read.py
@@ -16,6 +16,13 @@ from toolkit.clean.read_config import (
     resolve_clean_read_cfg,
 )
 from toolkit.clean.read_csv_normalized import _execute_normalized_csv_read
+from toolkit.clean.read_sql_utils import (
+    csv_trim_projection,
+    q_ident,
+    quote_list,
+    sql_path,
+)
+from toolkit.clean.read_excel import _execute_excel_read
 from toolkit.core.csv_read import (
     csv_read_option_strings,
     normalize_read_cfg,
@@ -43,31 +50,6 @@ class ReadInfo:
 
 
 
-
-
-def q_ident(name: str) -> str:
-    return '"' + name.replace('"', '""') + '"'
-
-
-def sql_path(p: Path) -> str:
-    s = p.resolve().as_posix()
-    return s.replace("'", "''")
-
-
-def quote_list(paths: list[Path]) -> str:
-    return ", ".join([f"'{sql_path(p)}'" for p in paths])
-
-
-def csv_trim_projection(columns: dict[str, str]) -> str:
-    exprs: list[str] = []
-    for name, dtype in columns.items():
-        qname = q_ident(name)
-        dtype_upper = dtype.upper()
-        if "CHAR" in dtype_upper or "TEXT" in dtype_upper or "STRING" in dtype_upper:
-            exprs.append(f"TRIM({qname}, ' \t\r\n') AS {qname}")
-        else:
-            exprs.append(qname)
-    return ", ".join(exprs)
 
 
 def _csv_read_options(
@@ -186,99 +168,6 @@ def _execute_parquet_read(
     return ReadInfo(source="parquet", params_used={})
 
 
-def _normalize_excel_sheet_name(value: Any) -> str | int:
-    if value is None:
-        return 0
-    if isinstance(value, bool):
-        raise ValueError("clean.read.sheet_name must be a string, integer, or null")
-    if isinstance(value, int):
-        return value
-    if isinstance(value, str):
-        text = value.strip()
-        if not text:
-            return 0
-        return text
-    raise ValueError("clean.read.sheet_name must be a string, integer, or null")
-
-
-def _trim_excel_dataframe(df: pd.DataFrame) -> pd.DataFrame:
-    return df.apply(
-        lambda column: column.map(lambda value: value.strip() if isinstance(value, str) else value)
-    )
-
-
-def _load_excel_frame(
-    input_file: Path,
-    read_cfg: dict[str, Any],
-) -> tuple[pd.DataFrame, dict[str, Any]]:
-    header = bool(read_cfg.get("header", True))
-    skip = int(read_cfg["skip"]) if read_cfg.get("skip") is not None else 0
-    trim_whitespace = read_cfg.get("trim_whitespace", True)
-    columns = read_cfg.get("columns")
-    sheet_name = _normalize_excel_sheet_name(read_cfg.get("sheet_name"))
-
-    df = pd.read_excel(
-        input_file,
-        sheet_name=sheet_name,
-        header=0 if header else None,
-        skiprows=skip,
-        dtype=object,
-        engine="openpyxl",
-    )
-
-    if columns:
-        expected_columns = list(columns.keys())
-        if len(expected_columns) != len(df.columns):
-            raise ValueError(
-                "Excel input columns mismatch. "
-                f"Configured={len(expected_columns)} detected={len(df.columns)} file={input_file}"
-            )
-        df.columns = expected_columns
-    elif not header:
-        df.columns = [f"col{i}" for i in range(len(df.columns))]
-
-    if trim_whitespace:
-        df = _trim_excel_dataframe(df)
-
-    return df, {
-        "sheet_name": sheet_name,
-        "header": header,
-        "skip": skip,
-        "trim_whitespace": bool(trim_whitespace),
-        "columns": dict(columns) if columns else None,
-    }
-
-
-def _execute_excel_read(
-    con: duckdb.DuckDBPyConnection,
-    input_files: list[Path],
-    read_cfg: dict[str, Any],
-    *,
-    logger,
-) -> ReadInfo:
-    frames: list[pd.DataFrame] = []
-    params_used: dict[str, Any] | None = None
-
-    for input_file in input_files:
-        frame, frame_params = _load_excel_frame(input_file, read_cfg)
-        frames.append(frame)
-        if params_used is None:
-            params_used = frame_params
-
-    combined = pd.concat(frames, ignore_index=True) if len(frames) > 1 else frames[0]
-    con.register("raw_input_df", combined)
-    con.execute("CREATE OR REPLACE VIEW raw_input AS SELECT * FROM raw_input_df;")
-
-    used = dict(params_used or {})
-    if used.get("columns") is None:
-        used.pop("columns", None)
-    logger.info(
-        "read_excel params used: source=excel params=%s",
-        json.dumps(used, ensure_ascii=False, sort_keys=True),
-    )
-    return ReadInfo(source="excel", params_used=used)
-
-
 def _validate_read_mode(mode: str) -> str:
     normalized_mode = str(mode or "fallback")
     if normalized_mode not in {"strict", "fallback", "robust"}:
@@ -389,7 +278,8 @@ def read_raw_to_relation(
         logger.info("read_csv params used: source=parquet params={}")
         return info
     if exts <= {".xlsx"}:
-        return _execute_excel_read(con, input_files, read_cfg, logger=logger)
+        result = _execute_excel_read(con, input_files, read_cfg, logger=logger)
+        return ReadInfo(source=result["source"], params_used=result["params_used"])
 
     normalized_mode = _validate_read_mode(mode)
     return _read_csv_relation(
